added reposts field  to post model and changed some logic at create and delete repost views
