+++ src/templates/post_app/my_posts.html
@@ -37,7 +37,7 @@
                 <button type="submit" class="btn btn-danger p-2">{{ post.liked.all.count }} likes</button>
             {% endif %}
         </form>
-        <a class="btn btn-light p-2 m-1">100 Reposts</a>
+        <a class="btn btn-light p-2 m-1">{{ post.reposts.all.count }} Reposts</a>
     </div>
 {% endblock %}
