+++ src/templates/post_app/user_posts.html
@@ -29,7 +29,7 @@
                 <button type="submit" class="btn btn-danger p-2 m-1">{{ post.liked.all.count }} likes</button>
             {% endif %}
         </form>
-        <a class="btn btn-light p-2 m-1">100 Reposts</a>
+        <a class="btn btn-light p-2 m-1">{{ post.reposts.all.count }} Reposts</a>
     </div>
 {% endblock %}
